Bestuur: enable searching posts by assigned names
Since bestuur positions are registered in post meta, they are by default
not included in post searches. This is different from when positions were
defined in the post content, which would be included in search queries.

This change adds posts to the search results that match one of their
assigned users or names for any related positions.

The search term(s) are matched first with related users through finding
the matched display names or login names. Then the posts are matched
based on post meta values for positions that either include a
corresponding user id or the searched term(s) itself.